Refactor WASM bindings into logical modules
- Split lib.rs into separate modules for better organization
- Create url_builder module for UrlBuilder and IndexRange types
- Create das_parser module for DAS parsing functionality
- Create converters module for type conversion utilities
- Maintain same public API through re-exports
